Fix demo-sw.js crash: move storage-touching deleteProject out of the shared handler
The demo service worker was crashing on load with:
  Uncaught TypeError: (0, Ts.promisify) is not a function
at a top-level `promisify(fs.readFile)` call.

Root cause: shared/handlers/projects.ts (imported by both the real server
and the demo build) had a runtime import of server/storage — specifically
getStorage() for deleteProject()'s cascade delete. server/storage/local.ts's
LocalStorageAdapter calls promisify(fs.readFile/writeFile/mkdir) eagerly at
module scope, which crashes immediately when that module gets pulled into
the demo service worker's bundle (no Node fs/util in a Worker global scope).
demo/api/router.ts imports other (portable) exports from the same file,
which drags in the whole module including its side-effecting imports.

Fix: extract the storage-touching deleteProject() into a new server-only
server/utils/delete-project.ts (matching the existing delete-run-files.ts
convention), importing the still-shared, DB-only deleteProjectData() from
shared/handlers/projects.ts. Demo mode already called deleteProjectData()
directly against its own in-browser DB and is unaffected. Updated the two
server API routes that called the old export.

Verified: rebuilt the demo bundle — the crash and the storage code are both
gone from the compiled demo-sw.js; loaded it in a browser and confirmed the
service worker activates and the app renders normally. Re-ran the mobile
regression suite, including its teardown (which calls the now-relocated
deleteProject via DELETE /api/tests/cleanup) — all 12 tests pass and
cleanup succeeds.
